requirements.txt
# python==3.7+
# vlc
# attention à la cohérence 64bits / 32bits
easygui==0.98.1
keyboard==0.13.4
pyliblo==0.10.0
python-vlc==3.0.7110
steppy==0.1.0
tk==0.1.0
